:truck: moved DefaultClient to Application.kt
